fix error with file writing, clean up printing, fix finding schema.yml

+++ utils/create_or_update_standard_models.py
@@ -8,6 +8,7 @@ import re
 from os import path
 import ruamel.yaml
 import sys
+from pprint import pprint
 
 dbt_string = """-- macro used to create this file can be found at:
 -- {github_path}
@@ -165,15 +166,17 @@ def create_or_update_docs(docs_path, destination_path):
     else:
         schema_dict['models'].append(docs_model)
         print(
-            f"Can we add docs for the current model to your schema.yml in {destination_path}?")
-        print(docs_model)
-        print("The resultant yaml would look like this:")
+            f"\nCan we add docs for the current model to your schema.yml in {destination_path}?")
+        print("This is the data we'd add to your schema.yml:")
+        pprint(docs_model)
+        print("\nThe resultant yaml would look like this:")
         yaml.dump(schema_dict, sys.stdout)
         choice = None
         while choice not in ['y', 'n']:
             choice = input("Type y for (y)es or n for (n)o:\n")
         if choice == 'y':
-            yaml.dump(schema_dict, schema_path)
+            with open(schema_path,'w') as f:
+                yaml.dump(schema_dict, f)
         return
     if docs_model == schema_dict['models'][existing_model_index]:
         return
@@ -184,16 +187,17 @@ def create_or_update_docs(docs_path, destination_path):
         print("Your existing in the schema file for this model")
         print("is different than the one in our documentation.")
         print("Here's what exists in our documentation but not your schema file:")
-        print(schema_dict_model_set-docs_model_set)
+        pprint(schema_dict_model_set-docs_model_set)
         print("Here's what exists in your schema file but not our documentation:")
-        print(docs_model_set-schema_dict_model_set)
+        pprint(docs_model_set-schema_dict_model_set)
         print("Would you like to update your schema file to match our documentation?")
         choice = None
         while choice not in ['y', 'n']:
             choice = input("Type y for (y)es or n for (n)o:\n")
         if choice == 'y':
             schema_dict['models'][existing_model_index] = docs_model
-            yaml.dump(schema_dict, schema_path)
+            with open(schema_path,'w') as f:
+                yaml.dump(schema_dict, f)
         return
 
 
@@ -268,7 +272,7 @@ def process_sources(sources_wanted, list_of_sources, macros_path, create, destin
                                 'macros', 'documentation')
                             docs_file = file.replace('sql', 'yml')
                             docs_file_path = path.join(docs_path, docs_file)
-                            create_or_update_docs(docs_path, destination_path)
+                            create_or_update_docs(docs_file_path, destination_path)
             for _, _, files in os.walk(destination_path):
                 for file in files:
                     destination_file_path = path.join(destination_path, file)
@@ -276,7 +280,8 @@ def process_sources(sources_wanted, list_of_sources, macros_path, create, destin
                     docs_path = source_path.replace('macros', 'documentation')
                     docs_file = file.replace('sql', 'yml')
                     docs_file_path = path.join(docs_path, docs_file)
-                    if (not os.path.exists(source_file_path)) and (not os.path.exists(docs_file_path)):
+                    if (not os.path.exists(source_file_path)) and (not os.path.exists(docs_file_path) and (
+                            file != 'schema.yml')):
                         delete_non_standard_model_choice(destination_file_path)
